# add caching for service catalog indexes