--- a/localstack/aws/spec.py
+++ b/localstack/aws/spec.py
@@ -35,18 +35,23 @@ def iterate_service_operations() -> Generator[Tuple[ServiceModel, OperationModel
             yield service, service.operation_model(op_name)
 
 
-class ServiceCatalog:
-    def get(self, name: ServiceName) -> Optional[ServiceModel]:
-        return self._services.get(name)
+@dataclasses.dataclass
+class ServiceCatalogIndex:
+    """
+    The ServiceCatalogIndex enables fast lookups for common operations to determine a service from service indicators.
+    """
 
-    def by_target_prefix(self, target_prefix: str) -> List[ServiceName]:
-        return self.target_prefix_index.get(target_prefix, [])
+    service_names: List[ServiceName]
+    target_prefix_index: Dict[str, List[ServiceName]]
+    signing_name_index: Dict[str, List[ServiceName]]
+    operations_index: Dict[str, List[ServiceName]]
+    endpoint_prefix_index: Dict[str, List[ServiceName]]
 
-    def by_signing_name(self, signing_name: str) -> List[ServiceName]:
-        return self.signing_name_index.get(signing_name, [])
 
-    def by_operation(self, operation_name: str) -> List[ServiceName]:
-        return self.operations_index.get(operation_name, [])
+class LazyServiceCatalogIndex:
+    """
+    A ServiceCatalogIndex that builds indexes in-memory from the spec.
+    """
 
     @cached_property
     def service_names(self) -> List[ServiceName]:
@@ -87,30 +92,14 @@ class ServiceCatalog:
 
     @cached_property
     def _services(self) -> Dict[ServiceName, ServiceModel]:
-        # since index creation requires all services to be loaded, we implement this method here
-        # to load and cache all services up front, and then let .get access this cache as well.
         return {service.service_name: service for service in list_services()}
 
 
-@dataclasses.dataclass
-class ServiceCatalogIndex:
-    service_names: List[ServiceName]
-    endpoint_prefix_index: Dict[str, List[ServiceName]]
-    operations_index: Dict[str, List[ServiceName]]
-    signing_name_index: Dict[str, List[ServiceName]]
-    target_prefix_index: Dict[str, List[ServiceName]]
-
-
-class CachedServiceCatalog(ServiceCatalog):
-    """
-    A ServiceCatalog that uses a pre-built ServiceCatalogIndex instead of resolving the indices
-    lazily at runtime from the specifications.
-    """
-
+class ServiceCatalog:
     index: ServiceCatalogIndex
 
-    def __init__(self, index: ServiceCatalogIndex):
-        self.index = index
+    def __init__(self, index: ServiceCatalogIndex = None):
+        self.index = index or LazyServiceCatalogIndex()
 
     @lru_cache(maxsize=512)
     def get(self, name: ServiceName) -> Optional[ServiceModel]:
@@ -136,26 +125,24 @@ class CachedServiceCatalog(ServiceCatalog):
     def endpoint_prefix_index(self) -> Dict[str, List[ServiceName]]:
         return self.index.endpoint_prefix_index
 
+    def by_target_prefix(self, target_prefix: str) -> List[ServiceName]:
+        return self.target_prefix_index.get(target_prefix, [])
 
-def save_service_index_cache(catalog: ServiceCatalog, file_path: str) -> ServiceCatalogIndex:
+    def by_signing_name(self, signing_name: str) -> List[ServiceName]:
+        return self.signing_name_index.get(signing_name, [])
+
+    def by_operation(self, operation_name: str) -> List[ServiceName]:
+        return self.operations_index.get(operation_name, [])
+
+
+def build_service_index_cache(file_path: str) -> ServiceCatalogIndex:
     """
-    Extracts from the given ServiceCatalog a ServiceCatalogIndex and pickles that into the given file.
-    :param catalog: the catalog that creates the service-operation indexes.
+    Creates a new ServiceCatalogIndex and stores it into the given file_path.
+
     :param file_path: the path to pickle to
     :return: the created ServiceCatalogIndex
     """
-    import pickle
-
-    cache = ServiceCatalogIndex(
-        catalog.service_names,
-        catalog.endpoint_prefix_index,
-        catalog.operations_index,
-        catalog.signing_name_index,
-        catalog.target_prefix_index,
-    )
-    with open(file_path, "wb") as fd:
-        pickle.dump(cache, fd)
-    return cache
+    return save_service_index_cache(LazyServiceCatalogIndex(), file_path)
 
 
 def load_service_index_cache(file: str) -> ServiceCatalogIndex:
@@ -169,3 +156,26 @@ def load_service_index_cache(file: str) -> ServiceCatalogIndex:
 
     with open(file, "rb") as fd:
         return pickle.load(fd)
+
+
+def save_service_index_cache(index: LazyServiceCatalogIndex, file_path: str) -> ServiceCatalogIndex:
+    """
+    Creates from the given LazyServiceCatalogIndex a ``ServiceCatalogIndex`, pickles its contents into the given file,
+    and then returns the newly created index.
+
+    :param index: the LazyServiceCatalogIndex to store the index from.
+    :param file_path: the path to pickle to
+    :return: the created ServiceCatalogIndex
+    """
+    import pickle
+
+    cache = ServiceCatalogIndex(
+        service_names=index.service_names,
+        endpoint_prefix_index=index.endpoint_prefix_index,
+        operations_index=index.operations_index,
+        signing_name_index=index.signing_name_index,
+        target_prefix_index=index.target_prefix_index,
+    )
+    with open(file_path, "wb") as fd:
+        pickle.dump(cache, fd)
+    return cache